enrich: fold categorize action in, unwrap wayback URLs
The categorize action duplicated enrich's fetch + worker pool +
file-walk + save logic to do nothing more than regex-match the source
HTML for category suggestions. Merged into enrich:

- New tool/categories.go holds the patterns and match/merge helpers
  (renamed from categorize.go; contents unchanged).
- mergeEnrichment now runs matchCategories over page.PlainText and
  unions any new hits into pm.Categories.
- Removed the categorize CLI action, the standalone CategorizePostmortems
  orchestrator, fetchBody, and printCategorizeReport.
- Renamed the leaked categorize* flag variables in main.go to enrich*
  (CLI flag names like -apply, -http-timeout are unchanged).

While here, fixed a related bug: when pm.URL is itself a Wayback
snapshot (https://web.archive.org/web/<ts>/<origin>) we now unwrap to
the origin and stash the snapshot in archive_url. The swap is
persisted even when the subsequent fetch fails, so a dead origin no
longer leaves the file pointing at the Wayback wrapper. Wayback
fetches also use the if_ flag so the page text is the snapshot
contents, not the Wayback chrome (which was leaking "Wayback Machine"
into title extraction).

Data: 21 entries had their url swapped to the original target with the
snapshot moved to archive_url. One malformed entry (basecamp 2018) had
markdown garbage in the YAML url field; cleaned by hand and the "see
also" reference moved into the body.
