+++ data/0b9c79f2-c9f6-4b88-a5b7-400973716d9a.md
@@ -1,6 +1,7 @@
 ---
 uuid: 0b9c79f2-c9f6-4b88-a5b7-400973716d9a
-url: https://web.archive.org/web/20220124104632/https://cercind.gov.in/2012/orders/Final_Report_Grid_Disturbance.pdf
+url: https://cercind.gov.in/2012/orders/Final_Report_Grid_Disturbance.pdf
+archive_url: https://web.archive.org/web/20220124104632if_/https://cercind.gov.in/2012/orders/Final_Report_Grid_Disturbance.pdf
 categories:
 - postmortem
 keywords:
@@ -26,4 +27,3 @@ Archive Team projects vary significantly in scale, from individual volunteers ar
 The text highlights the importance of these efforts for future generations and mentions specific initiatives like "Panic Downloads" for websites facing closure. It details how the Wayback Machine utilizes these collections to provide access to archived web content.
 
 However, the provided article text focuses exclusively on the operations of the Wayback Machine and the Archive Team. It does not contain any information regarding the specific incident referenced by the URL, such as its timeline, what failed, root cause, customer impact, or remediation efforts related to the 'Final Report Grid Disturbance.pdf'. Therefore, details pertaining to the incident itself cannot be extracted from this source.
-
