+++ data/118fcca7-cf5d-4b6d-924a-3cb6cf976f18.md
@@ -1,6 +1,7 @@
 ---
 uuid: 118fcca7-cf5d-4b6d-924a-3cb6cf976f18
-url: https://web.archive.org/web/20220704223244/https://flowdock-resources.s3.amazonaws.com/legal/Flowdock-RCA-For-Incident-On-2020-04-21.pdf
+url: https://flowdock-resources.s3.amazonaws.com/legal/Flowdock-RCA-For-Incident-On-2020-04-21.pdf
+archive_url: https://web.archive.org/web/20220704223244if_/https://flowdock-resources.s3.amazonaws.com/legal/Flowdock-RCA-For-Incident-On-2020-04-21.pdf
 categories:
 - cloud
 - postmortem
@@ -16,4 +17,3 @@ The provided article text is limited to Wayback Machine capture metadata and doe
 Information such as the precise timeline of the incident, the specific components that failed, the underlying root cause, and the full extent of customer impact cannot be determined from the given text. The remediation steps taken to resolve the issue are also not present.
 
 Consequently, a comprehensive description of the Flowdock incident cannot be generated based solely on the provided source material.
-
