+++ data/1c6f32bc-10ed-463a-a52d-763121f74e30.md
@@ -1,6 +1,7 @@
 ---
 uuid: 1c6f32bc-10ed-463a-a52d-763121f74e30
-url: https://web.archive.org/web/20211201033341/https://codeascraft.com/2012/01/23/solr-bittorrent-index-replication/
+url: https://codeascraft.com/2012/01/23/solr-bittorrent-index-replication/
+archive_url: https://web.archive.org/web/20211201033341if_/https://codeascraft.com/2012/01/23/solr-bittorrent-index-replication/
 title: Etsy site outage caused by multicast rsync
 categories:
 - cloud
@@ -32,4 +33,3 @@ Following the multicast rsync incident, Etsy decided to integrate BitTorrent int
 The BitTorrent replication process involves the primary server hashing index files, creating a torrent, and notifying replicas. Replicas then download the torrent file, verify existing index files, and use the BitTorrent protocol to download missing parts. This approach ensures index consistency through hash verification and provides a clearer overview of replication status compared to the previous HTTP method.
 
 While BitTorrent replication dramatically reduced transfer times (e.g., a 9 GB index to 23 replicas now takes less than an hour), it introduced a lag due to SHA1 hash calculations for large indexes. Etsy addressed this by creating a thread-safe SHA1 implementation and a DigestPool interface for parallel hashing, reducing the lag from 2 minutes to about 32 seconds. Future improvements include creating torrents per index segment to further reduce lag and supporting replication of arbitrary related files.
-
