+++ data/2552ccf5-4e03-40b1-a1fa-354a34bfa5fe.md
@@ -1,6 +1,7 @@
 ---
 uuid: 2552ccf5-4e03-40b1-a1fa-354a34bfa5fe
-url: https://web.archive.org/web/20220529044310/https://m.signalvnoise.com/postmortem-on-the-read-only-outage-of-basecamp-on-november-9th-2018/), [see also](https://web.archive.org/web/20220530044506/https://m.signalvnoise.com/update-on-basecamp-3-being-stuck-in-read-only-as-of-nov-8-922am-cst/
+url: https://m.signalvnoise.com/postmortem-on-the-read-only-outage-of-basecamp-on-november-9th-2018/
+archive_url: https://web.archive.org/web/20220529044310if_/https://m.signalvnoise.com/postmortem-on-the-read-only-outage-of-basecamp-on-november-9th-2018/
 categories:
 - postmortem
 company: Basecamp
@@ -8,4 +9,4 @@ product: ""
 
 ---
 
-In November 2018 a database hit the integer limit, leaving the service in read-only mode.
+In November 2018 a database hit the integer limit, leaving the service in read-only mode. See also the [follow-up on the read-only state](https://web.archive.org/web/20220530044506/https://m.signalvnoise.com/update-on-basecamp-3-being-stuck-in-read-only-as-of-nov-8-922am-cst/).
