+++ data/3aac3a45-0815-432e-8057-f8a02132cd36.md
@@ -1,6 +1,7 @@
 ---
 uuid: 3aac3a45-0815-432e-8057-f8a02132cd36
-url: http://web.archive.org/web/20170130231315/https://stackstatus.net/post/156407746074/outage-postmortem-january-24-2017
+url: https://stackstatus.net/post/156407746074/outage-postmortem-january-24-2017
+archive_url: https://web.archive.org/web/20170130231315if_/https://stackstatus.net/post/156407746074/outage-postmortem-january-24-2017
 title: Stack Exchange SQL Server bugcheck outage January 2017
 start_time: 2017-01-24T17:53:00Z
 end_time: 2017-01-24T18:10:00Z
@@ -32,4 +33,3 @@ The root cause involved a combination of the SQL Server bugcheck and a critical 
 Customer impact included the Stack Exchange network being inaccessible for 12 minutes after an initial 5-minute period of read-only access. Approximately 3.5 seconds of data may have been lost due to uncommitted transactions being rolled back.
 
 Remediation involved restarting the SQL service on NY-SQL02, which brought the network back online in read-only mode. After a sanity check, sites were restored to read-write functionality. NY-SQL02 has since been taken out of production for thorough testing, including memory diagnostics. Additionally, the SQL cluster was updated to 2016 SP1 CU1.
-
