+++ data/3d2adf6c-a887-4db4-badf-92987c613a46.md
@@ -1,6 +1,7 @@
 ---
 uuid: 3d2adf6c-a887-4db4-badf-92987c613a46
-url: https://web.archive.org/web/20220430011642/https://www.epicgames.com/fortnite/en-US/news/postmortem-of-service-outage-at-3-4m-ccu
+url: https://www.epicgames.com/fortnite/en-US/news/postmortem-of-service-outage-at-3-4m-ccu
+archive_url: https://web.archive.org/web/20220430011642if_/https://www.epicgames.com/fortnite/en-US/news/postmortem-of-service-outage-at-3-4m-ccu
 title: Fortnite service outages of February 3-4, 2018
 start_time: 2018-02-03T00:00:00Z
 end_time: 2018-02-05T00:15:00Z
@@ -34,4 +35,3 @@ The Account Service, responsible for user data and authentication, experienced a
 Beyond specific service issues, Epic Games also encountered general cloud capacity limits, specifically hitting the total instance limit in their AWS region, which affected their ability to scale services. Several API rate limits were also encountered. Furthermore, available IP exhaustion occurred in subnets, contributing to extended load balancer recovery times during incidents.
 
 Players experienced unusually long wait times, matchmaking instability, and were often blocked from signing in or out. Social features like seeing friends online or forming parties were largely non-functional, creating a "dark room" situation. In response, Epic Games is actively working on resolving database performance issues with external experts, optimizing backend calls, improving matchmaking data storage, enhancing operational excellence, and improving monitoring for cloud provider limits. Longer-term plans include rearchitecting the core messaging stack, expanding to multiple cloud providers and geographical locations to reduce blast radius, and scaling internal infrastructure.
-
