+++ data/55f106c4-1b4e-4d02-bae9-ccf07ff223c2.md
@@ -1,6 +1,7 @@
 ---
 uuid: 55f106c4-1b4e-4d02-bae9-ccf07ff223c2
-url: https://web.archive.org/web/20211019062735/https://www.pagerduty.com/blog/outage-post-mortem-april-13-2013/
+url: https://www.pagerduty.com/blog/outage-post-mortem-april-13-2013/
+archive_url: https://web.archive.org/web/20211019062735if_/https://www.pagerduty.com/blog/outage-post-mortem-april-13-2013/
 title: PagerDuty notification dispatch system outage of April 2013
 start_time: 2013-04-13T14:57:00Z
 end_time: 2013-04-13T16:23:00Z
@@ -36,4 +37,3 @@ The outage resulted in PagerDuty completely losing its ability to dispatch notif
 In the short term, PagerDuty implemented several improvements. These included adding more logging and aggregating logs for better searchability, planning to add a process watcher for automatic restarts of failed coordinator processes, and building a dashboard to enhance visibility into inter-host connectivity.
 
 For long-term remediation, PagerDuty committed to investing in staff training for Cassandra and ZooKeeper. They also planned to investigate moving off one of the affected AWS regions, emphasizing the need for thorough due diligence when selecting new hosting providers and datacenters to prevent similar single points of failure.
-
