+++ data/655ee350-a86f-424b-89c5-f80cc0fe527e.md
@@ -1,6 +1,7 @@
 ---
 uuid: 655ee350-a86f-424b-89c5-f80cc0fe527e
-url: https://web.archive.org/web/20201202234639/https://medium.com/@florian_7764/technical-post-mortem-of-the-august-incident-82ab4c3d6547
+url: https://medium.com/@florian_7764/technical-post-mortem-of-the-august-incident-82ab4c3d6547
+archive_url: https://web.archive.org/web/20201202234639if_/https://medium.com/@florian_7764/technical-post-mortem-of-the-august-incident-82ab4c3d6547
 title: Platform.sh EU region outage of August 2016
 start_time: 2016-08-18T19:00:00Z
 end_time: 2016-08-18T23:30:00Z
@@ -32,4 +33,3 @@ The incident occurred due to a series of cascading issues. During the orchestrat
 The orchestration software itself failed to start correctly due to connection drops to ZooKeeper. Investigation revealed a bug in the Kazoo library, which dropped connections when its internal pipe buffer became full (64k queries). After a fix for this was deployed, a subsequent issue was discovered: one ZooKeeper node exceeded its maximum allowed size, preventing full recovery.
 
 Remediation involved several steps. Automated checks were implemented to ensure the orchestration software is fully started before proceeding with further maintenance. A semaphore lock was added to the Kazoo client to prevent pipe buffer overflow, and a pull request is being prepared for the upstream Kazoo project. Finally, the ZooKeeper max buffer size was increased, and monitoring for ZooKeeper node sizes was implemented to prevent future occurrences.
-
