+++ data/922e216e-efe1-4687-a3e7-2398fbdd8dbe.md
@@ -1,6 +1,7 @@
 ---
 uuid: 922e216e-efe1-4687-a3e7-2398fbdd8dbe
-url: https://web.archive.org/web/20211124170124/https://medium.com/making-instapaper/instapaper-outage-cause-recovery-3c32a7e9cc5f
+url: https://medium.com/making-instapaper/instapaper-outage-cause-recovery-3c32a7e9cc5f
+archive_url: https://web.archive.org/web/20211124170124if_/https://medium.com/making-instapaper/instapaper-outage-cause-recovery-3c32a7e9cc5f
 title: Instapaper AWS RDS MySQL 2TB File Size Limit Outage
 start_time: 2017-02-09T20:30:00Z
 end_time: 2017-02-14T00:00:00Z
@@ -35,4 +36,3 @@ The incident resulted in 31 hours of full downtime, followed by five days of lim
 Recovery involved extensive collaboration with AWS support and Pinterest's Site Reliability Engineers. The primary path forward was to rebuild the 2.5TB database by performing a complete dump and restore to a new instance not subject to the 2TB limit. This was a multi-day effort. An Amazon engineer ultimately facilitated recovery by performing an rsync between the failed ext3 filesystem and a new ext4 filesystem, creating a new, fully indexed database. The new database was then synced with the temporary production database using binary logs, and finally promoted to master, restoring full service without data loss.
 
 The incident highlighted the lack of a disaster recovery plan for such a critical filesystem issue, leading to longer downtime and recovery times. Instapaper plans to define a better workflow for system-wide outages, escalating issues immediately to Pinterest's SRE team, and will increase backup testing frequency from quarterly to monthly. Despite the challenges, Instapaper will continue to use AWS RDS, acknowledging its overall reliability and the support received during the incident.
-
