+++ data/95f252f8-fef6-40a0-ab98-977f28d3dbba.md
@@ -1,6 +1,7 @@
 ---
 uuid: 95f252f8-fef6-40a0-ab98-977f28d3dbba
-url: http://web.archive.org/web/20150404235419/https://stackstatus.net/post/115305251014/outage-postmortem-march-31-2015
+url: https://stackstatus.net/post/115305251014/outage-postmortem-march-31-2015
+archive_url: https://web.archive.org/web/20150404235419if_/https://stackstatus.net/post/115305251014/outage-postmortem-march-31-2015
 title: Stack Exchange network outage due to StackEgg on March 31, 2015
 start_time: 2015-03-31T13:41:15Z
 end_time: 2015-03-31T13:47:00Z
@@ -31,4 +32,3 @@ Users experienced a 6-minute period where Stack Overflow was reported offline by
 Initial remediation involved raising the concurrent session limit on the HAProxy frontend and reducing StackEgg's request load. This was achieved by embedding necessary data directly into pages and increasing reliance on WebSockets for game data, significantly reducing HTTP requests.
 
 A subsequent traffic increase on April 1st prompted further action. The HTTP Keep-Alive duration for StackEgg traffic was temporarily lowered from 15 seconds to 5 seconds. This adjustment successfully reduced concurrent sessions from approximately 60,000 to around 19,000, stabilizing the network.
-
