+++ data/bb1d9c14-e8d8-4aa4-a806-31f31f3d96fe.md
@@ -1,6 +1,7 @@
 ---
 uuid: bb1d9c14-e8d8-4aa4-a806-31f31f3d96fe
-url: http://web.archive.org/web/20160720200842/https://stackstatus.net/post/147710624694/outage-postmortem-july-20-2016
+url: https://stackstatus.net/post/147710624694/outage-postmortem-july-20-2016
+archive_url: https://web.archive.org/web/20160720200842if_/https://stackstatus.net/post/147710624694/outage-postmortem-july-20-2016
 title: Stack Exchange Network outage of July 20, 2016
 start_time: 2016-07-20T14:44:00Z
 end_time: 2016-07-20T15:18:00Z
@@ -32,4 +33,3 @@ Since the homepage is used by the load balancer for health checks, its unrespons
 The root cause was a specific regular expression, `^[\s\u200c]+|[\s\u200c]+$`, designed to trim unicode whitespace. A malformed post containing approximately 20,000 consecutive whitespace characters exploited a performance characteristic of the backtracking regex engine. This caused the engine to perform an O(n²) number of character checks, leading to excessive CPU usage.
 
 To resolve the immediate issue, the problematic regular expression was replaced with a substring function. As follow-up actions, Stack Exchange plans to audit other regular expressions, improve post validation workflows, add controls to the load balancer to disable health checks during outages, and create a comprehensive outage checklist to streamline future incident responses.
-
