+++ data/c50f19f8-6289-405c-a459-1ac32dbd6955.md
@@ -1,6 +1,7 @@
 ---
 uuid: c50f19f8-6289-405c-a459-1ac32dbd6955
-url: http://web.archive.org/web/20160610080136/https://www.scribd.com/doc/309574927/ShapeShift-Post-Mortem-Public
+url: https://www.scribd.com/doc/309574927/ShapeShift-Post-Mortem-Public
+archive_url: https://web.archive.org/web/20160610080136if_/https://www.scribd.com/doc/309574927/ShapeShift-Post-Mortem-Public
 title: ShapeShift Cyberattack
 start_time: 2016-03-14T00:00:00Z
 end_time: 2016-04-09T10:23:08Z
@@ -33,4 +34,3 @@ The breaches resulted in the theft of cryptocurrency, as evidenced by coins bein
 In response, ShapeShift, guided by Ledger Labs Inc., undertook extensive corrective actions. These included replacing all employee computing hardware, communication channels, and cryptographic keys. The entire production environment was rebuilt using new cloud accounts, different operating systems, and automated deployment scripts with enhanced firewall rules and off-site logging.
 
 Furthermore, a bastion server was implemented to control access, and outbound firewall rules were tightened. New employee and infrastructure security policies were drafted and enforced. Future plans include implementing multi-signature architecture, deterministic keys, improved key backups, and a data sanitization policy to achieve higher CCSS compliance levels.
-
