+++ data/ca6ff7e2-3c95-4b2a-9b66-7a0b67fca25d.md
@@ -1,6 +1,7 @@
 ---
 uuid: ca6ff7e2-3c95-4b2a-9b66-7a0b67fca25d
-url: https://web.archive.org/web/20201101133510/https://keepthescore.co/blog/posts/deleting_the_production_database/
+url: https://keepthescore.co/blog/posts/deleting_the_production_database/
+archive_url: https://web.archive.org/web/20201101133510if_/https://keepthescore.co/blog/posts/deleting_the_production_database/
 title: Keepthescore.co production database accidental deletion
 start_time: 2020-10-17T21:45:00Z
 end_time: 2020-10-17T22:15:00Z
@@ -15,19 +16,25 @@ keywords:
 - keepthescore
 - backup
 - accidental
+- leaderboarded.com
+- python
+- flask
+- misconfiguration
 company: Keepthescore
 product: production database
 source_published_at: 2020-10-17T00:00:00+02:00
-source_fetched_at: 2026-05-04T17:53:17.472529Z
+source_fetched_at: 2026-05-04T18:18:19.548442Z
 summary: Engineers deleted the production database by accident. Database is a managed database from DigitalOcean with backups once a day. 30 minutes after the disaster, it went back online, however 7 hours of scoreboard data was gone forever.
 
 ---
 
-On Saturday, October 17, 2020, at approximately 10:45 PM CET, Keepthescore.co experienced a critical incident when its production database was accidentally deleted. The website was taken offline for 30 minutes for restoration efforts.
+On October 17, 2020, at approximately 10:45 PM CET, Leaderboarded.com experienced a critical incident when its production database was accidentally deleted. The service was restored by 11:15 PM CET, approximately 30 minutes later, by restoring from a daily backup.
 
-The incident occurred when a developer executed a function intended to delete and recreate a local development database. Despite the function's code explicitly setting the host to 'localhost', it inadvertently connected to and wiped the live production database. The exact mechanism that allowed this misconnection is still under investigation.
+The incident occurred when a developer executed a Python function intended to delete and recreate a local development database. Despite the function's code explicitly setting the host to 'localhost', it inadvertently connected to and wiped the live production database.
 
-This accidental deletion resulted in the loss of over 300,000 scoreboards and their associated data. Due to the daily backup schedule provided by DigitalOcean, any data created or updated between 3:47 PM CET and 11:21 PM CET on October 17, 2020, was permanently lost, as the restored backup predated this window.
+The root cause was identified as an incorrect environment variable setting (`FLASK_ENV=development`) which was not properly configured, leading the application to connect to the production environment instead of the development one.
 
-Keepthescore.co utilized DigitalOcean's managed database service, which facilitated a quick recovery from a daily backup. The service was restored by 11:15 PM CET, approximately 30 minutes after the incident began. The team plans to investigate the root cause of the production connection and implement measures to prevent a recurrence of this specific error.
+This accidental deletion resulted in the loss of over 300,000 scoreboards and their associated data. Any data created or updated between 3:47 PM CET and 11:21 PM CET on October 17, 2020, was permanently lost, as the restored backup predated this window.
+
+The author learned the importance of reliable backups, facilitated by DigitalOcean's managed database service, and the inherent danger of having database deletion functions readily available. Plans are in place to investigate the specific error and implement measures to prevent recurrence.
