+++ data/d44e5d9b-5c18-4cbe-b624-46bb9a311b84.md
@@ -1,6 +1,7 @@
 ---
 uuid: d44e5d9b-5c18-4cbe-b624-46bb9a311b84
-url: https://web.archive.org/web/20170728131458/https://kickstarter.engineering/the-day-the-replication-died-e543ba45f262
+url: https://kickstarter.engineering/the-day-the-replication-died-e543ba45f262
+archive_url: https://web.archive.org/web/20170728131458if_/https://kickstarter.engineering/the-day-the-replication-died-e543ba45f262
 title: Kickstarter MySQL replication failure
 start_time: 2013-03-07T00:00:00Z
 categories:
@@ -15,21 +16,24 @@ keywords:
 - bug
 - order by
 - kickstarter
+- data inconsistency
+- mixed mode replication
+- backer sequences
 company: Kickstarter
 product: Amazon RDS MySQL replication
 source_published_at: 2013-04-09T21:03:00Z
-source_fetched_at: 2026-05-04T17:53:47.511631Z
+source_fetched_at: 2026-05-04T18:18:49.845972Z
 summary: Primary DB became inconsistent with all replicas, which wasn't detected until a query failed. This was caused by a MySQL bug which sometimes caused `order by` to be ignored.
 
 ---
 
-On Thursday, March 7th, 2013, Kickstarter experienced a critical issue where all MySQL replicas stopped updating, leading to stale and inconsistent data across the site. The problem manifested as `Duplicate entry` errors during replication, indicating a divergence between the master and its replicas.
+On Thursday, March 7th, Kickstarter experienced a critical incident where all MySQL replicas stopped updating, leading to stale and inconsistent data across the site. The issue was signaled by `Duplicate entry` errors during replication, indicating a divergence between the master and its replicas.
 
 Investigation revealed that specific data, particularly backer sequences, had been inconsistent for days, with some inconsistencies dating back over a month. The replication process eventually broke when an "unsafe query" triggered row-based replication, exposing the underlying data divergence.
 
 The core issue was traced to a MySQL bug where the `ORDER BY` clause in an `UPDATE` query was sometimes ignored during replication. This, combined with InnoDB's transaction optimization causing transactions to finish out of order on the master, led to the master and replicas applying updates in different implicit orders, resulting in data inconsistency.
 
-The incident caused Kickstarter to scramble, requiring efforts to stabilize the site, minimize the effects of stale replicas, and communicate with users. The inconsistent backer sequences were crucial for creator reports. While the immediate crisis was resolved within hours, the underlying bug had caused data inconsistencies for an extended period.
+The incident caused Kickstarter to scramble, requiring efforts to stabilize the site, minimize the effects of stale replicas, and communicate with users. The inconsistent backer sequences were crucial for creator reports, impacting the accuracy of information provided to project creators.
 
-The primary outcome of the incident was the discovery and understanding of this critical MySQL replication bug. The team learned the importance of deeply understanding database internals beyond ORM abstractions. The article focuses on the discovery and analysis rather than a specific fix applied to the MySQL bug itself.
+While the immediate crisis was resolved within hours by stabilizing the site, the underlying bug had caused data inconsistencies for an extended period. The primary outcome of the incident was the discovery and understanding of this critical MySQL replication bug, highlighting the importance of deeply understanding database internals beyond ORM abstractions.
