+++ data/e16b28f3-b6d4-449b-ae93-cc8a4d074163.md
@@ -1,6 +1,7 @@
 ---
 uuid: e16b28f3-b6d4-449b-ae93-cc8a4d074163
-url: https://web.archive.org/web/20211016040522/https://blog.cloudflare.com/cloudflare-outage-on-july-17-2020/
+url: https://blog.cloudflare.com/cloudflare-outage-on-july-17-2020/
+archive_url: https://web.archive.org/web/20211016040522if_/https://blog.cloudflare.com/cloudflare-outage-on-july-17-2020/
 title: Cloudflare outage on July 17, 2020
 start_time: 2020-07-17T21:12:00Z
 end_time: 2020-07-17T21:39:00Z
@@ -17,21 +18,22 @@ keywords:
 - bgp
 - configuration error
 - router
+- backbone network
 company: Cloudflare
 product: Cloudflare backbone network
 source_published_at: 2020-07-18T01:22:00Z
-source_fetched_at: 2026-05-04T17:54:29.356701Z
+source_fetched_at: 2026-05-04T18:19:21.410703Z
 summary: During a maintenance of their private backbone network, an engineer made a typo in the Atlanta datacenter network configuration, causing all traffic coming from America and Europe flowing to this only datacenter, crushing it.
 
 ---
 
-On July 17, 2020, Cloudflare experienced an outage affecting its backbone network and various services globally. The incident began at 21:12 UTC and the primary service was restored by 21:39 UTC, though full recovery including logs and metrics took until 22:10 UTC. The outage caused traffic to drop by approximately 50% across Cloudflare's network, impacting numerous data centers in North America, Europe, and South America.
+On July 17, 2020, Cloudflare experienced a 27-minute outage affecting its backbone network and various services globally. The incident began at 21:12 UTC when a configuration change was made on a router in Atlanta. Service was largely restored by 21:39 UTC after the Atlanta router was disabled, with full recovery, including logs and metrics, achieved by 22:10 UTC.
 
-The root cause was a configuration error introduced during an attempt to alleviate congestion on the backbone link between Newark and Chicago. A network engineer made an incorrect change on a router in Atlanta. Instead of removing specific Atlanta routes from the backbone, a single line change inadvertently caused all BGP routes to be leaked into the backbone with a higher local-preference.
+The root cause was a configuration error introduced by a network engineer while attempting to alleviate congestion on a backbone link between Newark and Chicago. Instead of removing specific Atlanta routes, a single line change inadvertently caused all BGP routes to be leaked into the backbone with a higher local-preference (200) than local server routes (100).
 
-This misconfiguration instructed the Atlanta router to send all its BGP routes to other backbone routers, and because these routes had a higher local-preference (200) than the local server routes (100), all traffic intended for local compute nodes was redirected to Atlanta. This rapidly overwhelmed the Atlanta router, leading to failures in Cloudflare network locations connected to the backbone.
+This misconfiguration instructed the Atlanta router to send all its BGP routes to other backbone routers. Due to the higher local-preference, all traffic intended for local compute nodes was redirected to Atlanta, rapidly overwhelming the Atlanta router and causing Cloudflare network locations connected to the backbone to fail.
 
-The outage affected Internet properties and Cloudflare services, causing a significant drop in traffic. Affected locations included major cities across the US (San Jose, Dallas, Seattle, Los Angeles, Chicago, Washington, DC, Richmond, Newark, Atlanta), Europe (London, Amsterdam, Frankfurt, Paris, Stockholm, Moscow, St. Petersburg), and South America (São Paulo, Curitiba, Porto Alegre). Customers in these regions experienced service disruption for 27 minutes.
+The outage caused traffic to drop by approximately 50% across Cloudflare's network. Numerous data centers in North America (e.g., San Jose, Dallas, Chicago, Atlanta), Europe (e.g., London, Amsterdam, Frankfurt), and South America (e.g., São Paulo) were affected, leading to service disruption for customers in these regions.
 
-Cloudflare immediately disabled the overwhelmed Atlanta router to restore service. For long-term remediation, they deployed a change to prevent the recurrence of this specific issue by modifying the BGP local-preference for local server routes, ensuring a single location cannot attract other locations' traffic in a similar manner. They also planned to introduce a maximum-prefix limit on backbone BGP sessions to provide a fail-safe.
+Immediate remediation involved disabling the overwhelmed Atlanta router to restore service. For long-term prevention, Cloudflare deployed a change to modify the BGP local-preference for local server routes, ensuring a single location cannot attract other locations' traffic in a similar manner. They also planned to introduce a maximum-prefix limit on backbone BGP sessions as a fail-safe.
