+++ data/e3f52656-ec99-4b6e-a762-065f836ee370.md
@@ -1,6 +1,7 @@
 ---
 uuid: e3f52656-ec99-4b6e-a762-065f836ee370
-url: https://web.archive.org/web/20170202055452/https://www.businessinsider.com.au/a-fire-in-a-telstra-exchange-is-causing-flight-delays-and-network-outages-2017-2
+url: https://www.businessinsider.com.au/a-fire-in-a-telstra-exchange-is-causing-flight-delays-and-network-outages-2017-2
+archive_url: https://web.archive.org/web/20170202055452if_/https://www.businessinsider.com.au/a-fire-in-a-telstra-exchange-is-causing-flight-delays-and-network-outages-2017-2
 title: A fire in a Telstra exchange is causing flight delays and network outages
 start_time: 2017-02-02T03:37:00Z
 categories:
@@ -31,4 +32,3 @@ The disruption extended to critical services, with discount airline Jetstar expe
 Telstra confirmed the fire was contained but acknowledged the damage to its power infrastructure. Technicians were immediately dispatched to assess the full impact and begin restoration efforts.
 
 As a temporary measure, Telstra announced that SMS messages would be held while they worked to resolve the underlying issue. The company committed to providing more information as it became available and was working to resolve the situation as soon as possible.
-
