+++ data/edc911ec-e179-408a-b4f5-d3d6b0767069.md
@@ -1,6 +1,7 @@
 ---
 uuid: edc911ec-e179-408a-b4f5-d3d6b0767069
-url: https://web.archive.org/web/20210421222929/https://status.bintray.com/incidents/w4dfr0rpznkt
+url: https://status.bintray.com/incidents/w4dfr0rpznkt
+archive_url: https://web.archive.org/web/20210421222929if_/https://status.bintray.com/incidents/w4dfr0rpznkt
 title: Malicious packages reported in JCenter
 start_time: 2017-07-01T00:00:00Z
 end_time: 2018-12-12T08:00:00Z
@@ -31,4 +32,3 @@ The presence of malicious code within packages in JCenter meant that any users i
 Upon public disclosure in December 2018, Bintray immediately banned the fake user and removed all associated malicious packages and links from JCenter. Technical remediation included improving the observability of the reporting system with new monitoring alerts and logs.
 
 From a process and organizational perspective, the inclusion request process for the Support team was refreshed to specifically address impersonation attempts. Additionally, Bintray is establishing a dedicated moderation team to enhance inspection and validation of inclusion requests and authors, aiming for swift response to community moderation signals.
-
