+++ tool/categories_test.go
