+++ tool/main.go
@@ -27,20 +27,19 @@ import (
 )
 
 var (
-	log              = logging.Must(logging.NewLogger(postmortems.Service))
-	action           = flag.String("action", "", "")
-	dir              = flag.String("dir", "./data/", "")
-	categorizeApply  = flag.Bool("apply", false, "categorize/enrich: write changes back into the markdown files (default: dry run)")
-	categorizeWorker = flag.Int("workers", 8, "categorize: number of concurrent HTTP fetchers")
-	categorizeTime   = flag.Duration("http-timeout", 15*time.Second, "categorize/enrich: per-URL fetch timeout")
-	enrichOnly       = flag.String("only", "", "enrich: only process files whose name starts with this UUID prefix")
-	enrichForce      = flag.Bool("force", false, "enrich: overwrite non-empty fields (default: only fill blanks)")
-	enrichKeepDesc   = flag.Bool("keep-description", false, "enrich: preserve existing markdown body, only refresh metadata")
-	enrichMaxAge     = flag.Duration("max-age", 720*time.Hour, "enrich: skip files whose source_fetched_at is newer than this")
-	enrichWorkers    = flag.Int("enrich-workers", 4, "enrich: number of concurrent fetch+LLM workers")
-	gcpProject       = flag.String("gcp-project", os.Getenv("GOOGLE_CLOUD_PROJECT"), "enrich: GCP project for Vertex AI (defaults to GOOGLE_CLOUD_PROJECT)")
-	gcpLocation      = flag.String("gcp-location", "us-central1", "enrich: Vertex AI location/region")
-	geminiModel      = flag.String("gemini-model", "gemini-2.5-flash", "enrich: Gemini model name")
+	log            = logging.Must(logging.NewLogger(postmortems.Service))
+	action         = flag.String("action", "", "")
+	dir            = flag.String("dir", "./data/", "")
+	enrichApply    = flag.Bool("apply", false, "enrich: write changes back into the markdown files (default: dry run)")
+	enrichTimeout  = flag.Duration("http-timeout", 15*time.Second, "enrich: per-URL fetch timeout")
+	enrichOnly     = flag.String("only", "", "enrich: only process files whose name starts with this UUID prefix")
+	enrichForce    = flag.Bool("force", false, "enrich: overwrite non-empty fields (default: only fill blanks)")
+	enrichKeepDesc = flag.Bool("keep-description", false, "enrich: preserve existing markdown body, only refresh metadata")
+	enrichMaxAge   = flag.Duration("max-age", 720*time.Hour, "enrich: skip files whose source_fetched_at is newer than this")
+	enrichWorkers  = flag.Int("enrich-workers", 4, "enrich: number of concurrent fetch+LLM workers")
+	gcpProject     = flag.String("gcp-project", os.Getenv("GOOGLE_CLOUD_PROJECT"), "enrich: GCP project for Vertex AI (defaults to GOOGLE_CLOUD_PROJECT)")
+	gcpLocation    = flag.String("gcp-location", "us-central1", "enrich: Vertex AI location/region")
+	geminiModel    = flag.String("gemini-model", "gemini-2.5-flash", "enrich: Gemini model name")
 	qs               = []*survey.Question{
 		{
 			Name:     "url",
@@ -99,12 +98,11 @@ generate        Generate JSON files from the postmortem Markdown files.
 new             Create a new postmortem file.
 validate        Validate the postmortem files in the directory.
 serve           Serve the postmortem files in a small website.
-categorize      Scrape each postmortem URL and suggest additional categories.
-                Pass -apply to write suggestions back to the markdown files.
 enrich          Fetch each postmortem source URL (with Wayback fallback), extract metadata,
-                ask Gemini for incident times/product/expanded description, and write the
-                merged result back. Requires GOOGLE_APPLICATION_CREDENTIALS and a GCP
-                project. Pass -apply to write changes; -force to overwrite non-empty fields.
+                run regex-based category suggestions, ask Gemini for incident
+                times/product/expanded description, and write the merged result back.
+                Requires GOOGLE_APPLICATION_CREDENTIALS and a GCP project. Pass -apply to
+                write changes; -force to overwrite non-empty fields.
 `
 	danluuReadme = "https://raw.githubusercontent.com/danluu/post-mortems/master/README.md"
 	extractFile  = "./tmp/posts.md"
@@ -183,17 +181,6 @@ func main() {
 		err = newPostmortem(*dir)
 	case "validate":
 		_, err = postmortems.ValidateDir(*dir)
-	case "categorize":
-		var res []categorizeResult
-		res, err = CategorizePostmortems(categorizeOptions{
-			Dir:         *dir,
-			Apply:       *categorizeApply,
-			HTTPTimeout: *categorizeTime,
-			Concurrency: *categorizeWorker,
-		})
-		if err == nil {
-			printCategorizeReport(os.Stdout, res, *categorizeApply)
-		}
 	case "enrich":
 		err = runEnrich()
 	case "serve":
@@ -223,11 +210,11 @@ func runEnrich() error {
 	res, err := EnrichPostmortems(ctx, enrichOptions{
 		Dir:             *dir,
 		Only:            *enrichOnly,
-		Apply:           *categorizeApply,
+		Apply:           *enrichApply,
 		Force:           *enrichForce,
 		KeepDescription: *enrichKeepDesc,
 		MaxAge:          *enrichMaxAge,
-		HTTPTimeout:     *categorizeTime,
+		HTTPTimeout:     *enrichTimeout,
 		Concurrency:     *enrichWorkers,
 		LLM:             llm,
 		Logger:          logger,
@@ -235,7 +222,7 @@ func runEnrich() error {
 	if err != nil {
 		return err
 	}
-	LogEnrichReport(logger, res, *categorizeApply)
+	LogEnrichReport(logger, res, *enrichApply)
 	return nil
 }
